+++ container/agent-runner/src/scheduling/task-script.ts
@@ -1,8 +1,8 @@
 import { execFile } from 'node:child_process';
 import fs from 'node:fs';
 import path from 'node:path';
-import type { MessageInRow } from './db/messages-in.js';
-import { touchHeartbeat } from './db/connection.js';
+import type { MessageInRow } from '../db/messages-in.js';
+import { touchHeartbeat } from '../db/connection.js';
 
 const SCRIPT_TIMEOUT_MS = 30_000;
 const SCRIPT_MAX_BUFFER = 1024 * 1024;
